refactor: move ResizedVideo and MyUCF101 to TubeViT
