+++ TubeViT/video_transforms.py
@@ -1,0 +1,33 @@
+import torch
+from torchvision.transforms._functional_video import resize
+
+
+class ResizedVideo:
+    def __init__(
+        self,
+        size,
+        interpolation_mode="bilinear",
+    ):
+        if isinstance(size, tuple):
+            assert len(size) == 2, "size should be tuple (height, width)"
+            self.size = size
+        else:
+            self.size = (size, size)
+
+        self.interpolation_mode = interpolation_mode
+
+    def __call__(self, clip: torch.Tensor):
+        """
+        Args:
+            clip (torch.tensor): Video clip to be cropped. Size is (C, T, H, W)
+        Returns:
+            torch.tensor: resized video clip.
+                size is (C, T, H, W)
+        """
+        return resize(clip, self.size, self.interpolation_mode)
+
+    def __repr__(self):
+        return self.__class__.__name__ + \
+            '(size={0}, interpolation_mode={1})'.format(
+                self.size, self.interpolation_mode
+            )
