+++ scripts/train.py
@@ -1,66 +1,17 @@
 import os
 import pickle
-from typing import Tuple, Optional, Callable
 
 import click
 import matplotlib.pyplot as plt
 import pytorch_lightning as pl
-import torch
-from torch import Tensor
 from torch.utils.data import DataLoader, RandomSampler
-from torchvision.datasets import UCF101
 from torchvision.transforms import transforms as T
-from torchvision.transforms._functional_video import resize
 from torchvision.transforms._transforms_video import RandomResizedCropVideo, RandomHorizontalFlipVideo, ToTensorVideo, \
     NormalizeVideo
 
+from TubeViT.dataset import MyUCF101
 from TubeViT.model import TubeViTLightningModule
-
-
-class ResizedVideo:
-    def __init__(
-        self,
-        size,
-        interpolation_mode="bilinear",
-    ):
-        if isinstance(size, tuple):
-            assert len(size) == 2, "size should be tuple (height, width)"
-            self.size = size
-        else:
-            self.size = (size, size)
-
-        self.interpolation_mode = interpolation_mode
-
-    def __call__(self, clip: torch.Tensor):
-        """
-        Args:
-            clip (torch.tensor): Video clip to be cropped. Size is (C, T, H, W)
-        Returns:
-            torch.tensor: resized video clip.
-                size is (C, T, H, W)
-        """
-        return resize(clip, self.size, self.interpolation_mode)
-
-    def __repr__(self):
-        return self.__class__.__name__ + \
-            '(size={0}, interpolation_mode={1})'.format(
-                self.size, self.interpolation_mode
-            )
-
-
-class MyUCF101(UCF101):
-    def __init__(self, transform: Optional[Callable] = None, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        self.transform = transform
-
-    def __getitem__(self, idx: int) -> Tuple[Tensor, int]:
-        video, audio, info, video_idx = self.video_clips.get_clip(idx)
-        label = self.samples[self.indices[video_idx]][1]
-
-        if self.transform is not None:
-            video = self.transform(video)
-
-        return video, label
+from TubeViT.video_transforms import ResizedVideo
 
 
 @click.command()
@@ -70,7 +21,7 @@ class MyUCF101(UCF101):
 @click.option('-b', '--batch-size', type=int, default=32, help='batch size.')
 @click.option('-f', '--frames-per-clip', type=int, default=32, help='frame per clip.')
 @click.option('-v', '--video-size', type=click.Tuple([int, int]), default=(224, 224), help='frame per clip.')
-@click.option('--max-epochs', type=int, default=5, help='max epochs.')
+@click.option('--max-epochs', type=int, default=10, help='max epochs.')
 @click.option('--num-workers', type=int, default=0)
 @click.option('--fast-dev-run', type=bool, is_flag=True, show_default=True, default=False)
 @click.option('--seed', type=int, default=42, help='random seed.')
@@ -177,7 +128,14 @@ def main(dataset_root, annotation_path, num_classes, batch_size, frames_per_clip
         weight_path='tubevit_b_(a+iv)+(d+v)+(e+iv)+(f+v).pt',
     )
 
-    trainer = pl.Trainer(max_epochs=max_epochs, accelerator='auto', fast_dev_run=fast_dev_run)
+    callbacks = [pl.callbacks.LearningRateMonitor(logging_interval='epoch')]
+
+    trainer = pl.Trainer(
+        max_epochs=max_epochs,
+        accelerator='auto',
+        fast_dev_run=fast_dev_run,
+        callbacks=callbacks,
+    )
     trainer.fit(model, train_dataloaders=train_dataloader, val_dataloaders=val_dataloader)
     trainer.save_checkpoint('./models/tubevit_ucf101.ckpt')
